Dock
This code allows a player or NPC object to move into a certain distance of a dock, and pop out the other side. If entering as a human, they will pop out the other side of the dock as a ship. If entering as a ship, they will pop out the other side as a human.