Create SceneManager
This code enables entire scenes to be saved into a stack. This allows the player to enter a city from an overview map. The player can do whatever they want inside the city, and when they choose to leave the overview map will have all assets in the same exact location as when the player entered the city. This can stack any number of times. So the player could enter a city from the overview map, then go into the city sewers, then enter combat. As the player leaves each scene, the scene they originally migrated from will be exactly how they left it.